requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.23.2
pdf2docx==0.6.4
pdf2image==1.16.3
Pillow==10.3.0
python-docx==0.8.12
reportlab==4.1.3
